Fix liveness bug in aggregator error handling
In normal flow, tryBuildFinalProof disables further proof building
until sendFinalProof re-enables it, to ensure that there is only
one proof in flight at a time. However, if tryBuildFinalProof fails
to build a proof for any reason, then sendFinalProof will not be
called, so it is tryBuildFinalProof's responsibility to re-enable
proof building in this case.

The defer statement in tryBuildFinalProof re-enables proof building
in the case where tryBuildFinalProof fails to build a proof and
returns an error. However, there are a few cases where
tryBuildFinalProof will fail to build a proof and _not_ return an
error (for example, if no eligible proof is ready). The defer
statement does not catch this case, so proof building remains
disabled, and furthermore, future calls to tryBuildFinalProof exit
early _because_ proof building is disabled, and so neither
tryBuildFinalProof nor sendFinalProof ever re-enable proof building.

This changes the defer statement to check the flag indicating
whether a proof was generated instead of checking the error variable,
so that it will re-enable proof building in _all_ cases where a
proof was not generated.

This should fix the issue where integration tests for the sequencer
hang because they never receive a state transition proof. It also
allows us to shorten the proof delay time for testing purposes,
whereas doing this before would trigger the error paths that this
change addresses.