Simplify GetPreconfirmations
Since we are no longer combining the events returned by
GetPreconfirmations into the block order returned by
GetRollupInfoByBlockRange, we can build the block order in a much
simpler way: we don't have to worry about coalescing with
existing blocks. In fact, the procedure for appending sequenced
batches is now the same as in newBlocksEvent, and has been
factored out into a helper function, appendSequencedBatches.