Update Geth to v1.12
This is the version used by the upstream, and contains a new
block header field (`WithdrawalsHash`) which is set by Sepolia,
which may be important when computing hashes of blocks returned by
a Sepolia RPC.

Lockfile changes are from manually updating go-ethereum to v1.12.0
with `go get`, and then running `go mod tidy`. Code changes are
based on upstream, but taken manually rather than trying to merge
or cherry-pick.