Fail in processSequenceBatches if we don't have the parent batch
This will cause an error if there has been a reorg since we fetched the
batches. In normal mode, this should never happen, since we check for
reorgs and then fetch batches synchronously. But in preconfirmations
mode, the batches are fetched asynchronously wrt to the reorg
detection, and so the L2 head can change under us. In this case the
new check will fail and we will re-fetch preconfirmations after
checking the new L2 head.

Fixes EspressoSystems/espresso-polygon-zkevm-demo#246