Decouple batch L1 origin from L1 block where batch was received
This allows us to use the L1 head from HotShot as the L1 block
number when reading batches from NewBlocks events in regular mode,
without messing up the rest of the synchronizer which assumes the
L1 block number is the current L1 block we're synchronizing. This
brings the regular node into consistency with the preconfirmations
node.