Detect reorgs during event processing
Check the hash of the retrieved L1 block against the block hash in
the event log we processed. This will detect if the block has changed
(due to a reorg) in the time since we processed it. In this case we
simply return an error so that the synchronizer will retry fetching
the events, now with the new L1 state.

If we don't check this here, we might never detect this reorg, because
we will store the hash of the retrieved block (from after the reorg)
even though we processed events from the block which got reorged out.
The stored hash will then match the hash when we query it again from
the L1 in checkReorg, and we will never process events from the new
block.